projects/project_60/requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
arrow==0.17.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.6.0
cookiecutter==1.7.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Fiona==1.8.17
geopandas==0.8.1
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kaleido==0.0.3.post1
kiwisolver==1.2.0
libpysal==4.3.0
MarkupSafe==1.1.1
matplotlib==3.3.2
Mesa==0.8.7
mesa-geo==0.2.0
mistune==0.8.4
munch==2.5.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
plotly==4.11.0
poyo==0.5.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
retrying==1.3.3
Rtree==0.9.4
scipy==1.5.3
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
soupsieve==2.0.1
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.3.1
